refactor: AboutScreen, FeebackScreen & SettingsScreen
